Added immediate_or_start combinator.
