+++ swim_utilities/swim_future/src/combinators/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod immediate_or_join;
+mod immediate_or;
 #[cfg(test)]
 mod tests;
 
@@ -26,7 +26,9 @@ use std::pin::Pin;
 use std::sync::Arc;
 use tokio::sync::Notify;
 
-pub use immediate_or_join::{immediate_or_join, ImmediateOrJoin};
+pub use immediate_or::{
+    immediate_or_join, immediate_or_start, ImmediateOrJoin, ImmediateOrStart, SecondaryResult,
+};
 
 /// A future that transforms another future using [`Into`].
 #[pin_project]
